subsystem_restart: do not use WQ_CPU_INTENSIVE in an unbound workqueue allocation
As per documentation [1], WQ_CPU_INTENSIVE is useful for bound work items which
are expected to hog CPU cycles so that their execution is regulated by the system
scheduler and thereby do not contribute to the concurrency level which is pointless
for an unbound workqueue.
Therefore, using WQ_CPU_INTENSIVE with WQ_BOUND is redundant. Remove it.

[1]: https://www.kernel.org/doc/html/latest/core-api/workqueue.html#flags

Change-Id: b19fe934c256c283f1a558df8c5d4a0b65afd70c
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>